Creates the View Artist Show Page
Within the routes.rb file I changed the default params for the artists
routes to become "artist_name". From there I created the "show" method for the
ArtistsController.rb file that then led for the possibility of any user
to be able to access a specific artist's page that shows the albums our website
has in stock.