adds user creation
Added the ability for a guest to create an account. That includes
after creation sending them to their personal dashboard. The route
was altered to show `/dashboard` instead of `/users/:id`. Also
created some view helpers that check if there is a current user and
renders specific views for login / logout and logged in status. Added
bcrypt (has_secure_password) to the user model.