Help Needed!
And by help I'm asking for Julian's help.
I'm stuck in figuring out a way on how to display the total quantity per
individual albums. This wouldn't be that big of an issue but since the iteration
over the @cart.albums doesn't allow me to call on its specific quantity that is
where I'm stuck at.