Added the Plus & Minus Quantity Functionality
Within this pull request you will find that I managed to figure a way out in
subtracting the quantity of albums within the cart page. I also made sure that
whenever the album hits zero quantity that the album no longer shows in the
cart's page. Other than that I do believe this still needs some refactoring to
do since we should avoid putting in a lot of logic within our views.

This also includes the work that Scott and Admir did which was include the plus
functionality for the albums. One final thing worth noting is that within the
cart's index view Julian and I ended up adding an operator key that
differentiates between the two put's request within the params for our update
method.