Dynamic Cart Total for Navigation
Within this pull request you will find that whenever a user adds an album to
their cart that it will display as so on the nav bar, "Cart (albums_quantity)".
I also made sure to add the logic necessary to make sure if a user doesn't have
anything in their cart that it will not display "Cart (0)" but rather "Cart".