An Order's Detail
On Order's Controller when an order is successful, it will flash a message
confirming that the order was successful while on a redirected orders show
path. I've hit a plateau since we will need to implement the SimpleDelegator
onto the new file I created called album_retriever.rb. This is an issue since
on the issue's user story I need to see specific detailed information of an item
within the order.