Admin can edit an album
The "edit" links on the admin albums index view now point to an edit album page.
Once the album is edited, the admin is returned to the admin albums index.
If save isn't successful, the edit page will re-render.
I had to revise another test due to factory girl weirdness.